added core tests examples
