+++ .gitignore
@@ -3,6 +3,7 @@ venv/
 __pycache__/
 .pytest_cache/
 .mypy_cache/
+sandbox/
 
 # Django stuff:
 db.sqlite3
