+++ pytest.ini
@@ -1,10 +1,8 @@
 [pytest]
+DJANGO_SETTINGS_MODULE = CodeReviewBot.CodeReviewBot.settings
 python_files = tests.py test_*.py *_tests.py
 
-log_cli = 1
-log_cli_level = WARNING
-log_cli_format = %(message)s
-
+log_cli = 0
 
 log_file = pytest.log
 log_file_level = DEBUG
