+++ setup.cfg
@@ -16,6 +16,7 @@ statistics = True
 
 [mypy]
 check_untyped_defs = false
+files = CodeReviewBot,tests
 
 [mypy-django.*]
 ignore_missing_imports = True
