+++ tests/core/conftest.py
@@ -1,16 +1,20 @@
-import pytest
 import logging
+
+import pytest
 
 
 @pytest.fixture
 def foo():
-    logging.warning('a warning')
     return 1
 
 
 @pytest.fixture(scope='session')
 def bar():
-    logging.error('an error')
     return 2
 
 
+@pytest.fixture
+def baz():
+    logging.info('baz setup')
+    yield 3
+    logging.info('baz teardown')
