+++ tests/core/test_basic_batabase.py
@@ -1,0 +1,8 @@
+import pytest
+from django.contrib.auth.models import User
+
+
+@pytest.mark.django_db
+def test_elementary_db():
+    User.objects.create_user('john', 'lennon@thebeatles.com', 'johnpassword')
+    assert User.objects.count() == 1
