--- tests/core/test_somefile.py
@@ -1,8 +1,0 @@
-def test_f_one(foo):
-    x = foo
-    assert x == 1
-
-
-def test_b_one(bar):
-    assert True
-
